+++ eneeyes-webapp/src/main/java/br/com/eneeyes/main/model/PositionAlarm.java
@@ -1,5 +1,6 @@
 package br.com.eneeyes.main.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.CascadeType;
@@ -51,7 +52,7 @@ public class PositionAlarm {
 	private Date lastUpdate;
 
 	@Column(name = "LAST_VALUE", nullable = true)
-	private Double lastValue;
+	private BigDecimal lastValue;
     		
 	@ManyToOne(cascade=CascadeType.DETACH, fetch = FetchType.LAZY)
 	@JoinColumn(name="COMPANY_DETECTOR_ID", nullable = false)
@@ -120,11 +121,11 @@ public class PositionAlarm {
 		this.lastUpdate = lastUpdate;
 	}
 
-	public Double getLastValue() {
+	public BigDecimal getLastValue() {
 		return lastValue;
 	}
 
-	public void setLastValue(Double lastValue) {
+	public void setLastValue(BigDecimal lastValue) {
 		this.lastValue = lastValue;
 	}
